+++ src/client.rs
@@ -15,7 +15,7 @@ use crate::{
     shared::{
         backend::channels::{ClientChannel, ServerChannel},
         replication::{
-            deferred_entity::{DeferredChanges, DeferredEntity},
+            deferred_entity::{DeferredEntity, EntityScratch},
             mutate_index::MutateIndex,
             receive_markers::{EntityMarkers, ReceiveMarkers},
             registry::{
@@ -128,7 +128,7 @@ impl Plugin for ClientPlugin {
 /// See also [`ReplicationMessages`](crate::server::replication_messages::ReplicationMessages).
 pub(super) fn receive_replication(
     world: &mut World,
-    mut changes: Local<DeferredChanges>,
+    mut scratch: Local<EntityScratch>,
     mut entity_markers: Local<EntityMarkers>,
     mut entity_buffer: Local<EntityBuffer>,
 ) {
@@ -150,7 +150,7 @@ pub(super) fn receive_replication(
     let mut mutate_ticks = world.remove_resource::<ServerMutateTicks>();
 
     let mut params = ReceiveParams {
-        changes: &mut changes,
+        scratch: &mut scratch,
         entity_markers: &mut entity_markers,
         entity_buffer: &mut entity_buffer,
         entity_map: &mut entity_map,
@@ -234,6 +234,9 @@ fn apply_replication(
     for mut message in messages.receive(ServerChannel::Updates) {
         if let Err(e) = apply_update_message(world, params, &mut message) {
             error!("unable to apply update message: {e}");
+
+            // SAFETY: components in the scratch were pushed using this world.
+            unsafe { params.scratch.manual_drop(world.components()) };
             params.entity_buffer.free(world);
         }
     }
@@ -400,6 +403,9 @@ fn apply_mutate_messages(
                     "unable to apply mutate message for tick `{:?}`: {e}",
                     mutate.message_tick
                 );
+
+                // SAFETY: components in the scratch were pushed using this world.
+                unsafe { params.scratch.manual_drop(world.components()) };
                 params.entity_buffer.free(world);
             }
         }
@@ -486,7 +492,7 @@ fn apply_removals(
 
     let Ok(mut client_entity) = world
         .get_entity_mut(client_entity)
-        .map(|entity| DeferredEntity::new(entity, params.changes))
+        .map(|entity| DeferredEntity::new(entity, params.scratch))
     else {
         // Client could predict despawn.
         debug!("ignoring removals for despawned `{client_entity}`");
@@ -538,7 +544,7 @@ fn apply_changes(
     let data_size: usize = postcard_utils::from_buf(message)?;
 
     let world_cell = world.as_unsafe_world_cell();
-    let entity_allocator = world_cell.entities_allocator();
+    let entity_allocator = world_cell.entity_allocator();
     // SAFETY: used only to create `DeferredEntity`, which won't let mutably alias `EntityAllocator`.
     let world = unsafe { world_cell.world_mut() };
 
@@ -551,7 +557,7 @@ fn apply_changes(
                 return Ok(());
             };
 
-            let mut client_entity = DeferredEntity::new(client_entity, params.changes);
+            let mut client_entity = DeferredEntity::new(client_entity, params.scratch);
             if !client_entity.contains::<Remote>() {
                 // Even though the entity already exists, it could have been spawned during
                 // deserialization of another component and doesn't have the marker yet.
@@ -560,7 +566,7 @@ fn apply_changes(
             client_entity
         }
         EntityEntry::Vacant(entry) => {
-            let mut client_entity = DeferredEntity::new(world.spawn_empty(), params.changes);
+            let mut client_entity = DeferredEntity::new(world.spawn_empty(), params.scratch);
             client_entity.insert(Remote);
             entry.insert(client_entity.id());
             client_entity
@@ -680,13 +686,13 @@ fn apply_mutations(
     };
 
     let world_cell = world.as_unsafe_world_cell();
-    let entity_allocator = world_cell.entities_allocator();
+    let entity_allocator = world_cell.entity_allocator();
     // SAFETY: used only to create `DeferredEntity`, which won't let mutably alias `EntityAllocator`.
     let world = unsafe { world_cell.world_mut() };
 
     let Ok(mut client_entity) = world
         .get_entity_mut(client_entity)
-        .map(|entity| DeferredEntity::new(entity, params.changes))
+        .map(|entity| DeferredEntity::new(entity, params.scratch))
     else {
         // Client could predict despawn.
         debug!("ignoring mutations for despawned `{client_entity}`");
@@ -785,7 +791,7 @@ fn apply_mutations(
 ///
 /// To avoid passing a lot of arguments into all receive functions.
 struct ReceiveParams<'a> {
-    changes: &'a mut DeferredChanges,
+    scratch: &'a mut EntityScratch,
     entity_markers: &'a mut EntityMarkers,
     entity_buffer: &'a mut EntityBuffer,
     entity_map: &'a mut ServerEntityMap,