+++ tests/spawn.rs
@@ -184,6 +184,31 @@ fn related() {
 }
 
 #[test]
+fn resource() {
+    let mut server_app = App::new();
+    let mut client_app = App::new();
+    for app in [&mut server_app, &mut client_app] {
+        app.add_plugins((
+            MinimalPlugins,
+            StatesPlugin,
+            RepliconPlugins.set(ServerPlugin::new(PostUpdate)),
+        ))
+        .replicate_resource::<R>()
+        .finish();
+    }
+
+    server_app.connect_client(&mut client_app);
+
+    server_app.world_mut().insert_resource(R);
+
+    server_app.update();
+    server_app.exchange_with_client(&mut client_app);
+    client_app.update();
+
+    assert!(client_app.world().contains_resource::<R>());
+}
+
+#[test]
 fn empty_before_connection() {
     let mut server_app = App::new();
     let mut client_app = App::new();
@@ -603,6 +628,9 @@ struct A;
 
 #[derive(Component, Deserialize, Serialize)]
 struct B;
+
+#[derive(Resource, Deserialize, Serialize)]
+struct R;
 
 #[derive(Component)]
 #[component(immutable)]